Add split and mv data_source to dataset
